Some cleaning. Extracted query methods to new class Query.cs.
